Add more example calls
This also refactors the flow a bit towards get methods.
