+++ nodejs/src/credentials.js
